Introduced optional safe margin when clearing the buffer to avoid rebufferings.
A new optional parameter called safeMargin is introduced to avoid clearing the buffer entirely. The value of such offset should be computed in a custom ABR manager and provided to the player and the streaming engine. Since the buffer could be busy when clearBuffer is called, another parameter had to be introduced to save the value of the safe margin when the call to clear the buffer is deferred to a later update.
When a safe margin value is provided, the clearBuffer method uses ‘remove’ as method to clear the buffer, and not the method ‘clear’.